requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.16
alembic==1.10.2
python-multipart==0.0.6
passlib==1.7.4
requests==2.31.0
dataclasses-json==0.5.7 
python-jose==3.3.0  
python-dotenv
pydantic[email]
sqlalchemy-utils
nanoid
pytz
itsdangerous
asyncpg
psycopg2
cryptography 